Feature Update - Update Checks!
Added the ability to turn on update checks. If you enable this feature,
the script will parse the dev RSS feed for PATCH titles. If a new one is
found it will automatically restart the server to apply the update. You
need to have SteamCMD enabled for this to work properly. Also corrected
some bugs.