Restart at different Times throught the day
Had a request to allow multiple restarts through the day. This makes
that possible. There are 4 hour options available now. If you turn on
daily restarts and set them all to the same hour it will restart once a
day. If you set them to different hours it will restart at each hour.