Feature Addition
Adds two options to the INI file. One is to disable using MULTIHOME to
set the IP. Set BindIP=no if you are having connection issues.  Also,
adds an option to close the windows that come up when running the server
with steam open. This will automatically close them and continue to
start the server. Set SteamFix=yes