Added Log Rotate with Options
Added the ability to turn on log rotate. Set number of logs required and
how often in hours you would like to rotate them. Also condensed all log
writes to one log file and modified the way the date and time was
written to the log. Finally, made modifications to the Update check to
allow intervals from 5 minutes to forever.