Closes #1
Moved functions out from main script body to make navigation of main
code easier.  Moved INI to it's own script as well due to the size of
the functions.  Modified the INI handling to check for all variables in
user INI. If any are found missing, it saves all variables that were
found, backs up the current INI, and writes a new INI with the users
existing variables plus the defaults for the missing ones.