Clean Up Update Check
This closes #21  . It will check for any title with PATCH in it. It
would still be possible to false update if a title had the word patch in
it. Shouldn't miss a patch now though. As long as they post PATCH in the
title it will find the first title with that word, check it against the
full title from before, and then restart if they don't match.  Removed
the excess logging.