Fix Possible Conflict when running multiple Servers
This fixes a possible conflict when running multiple servers on the same
machine. Ideally, you would use two seperate SteamCMD directories, but
if you don't this will look for an existing file during update checks.
If it finds it, it will wait for 1 minute to allow the other script time
to update. After that it will delete the file and run the update check.