Start Adding Enable/Disable Settings by Time
Added functions for enabling and disabling settings by time. Added
functions to call the enable/disable checks for enabling building damage
and disabling avatars on a set schedule.  Began making changes to
multiple variables, and trying to clean up code some.  This is a
potentially breaking update, and should be treated as such until fully
tested.