Create Option to Flip Schedule
This update will allow the option of flipping the schedules for eithr
Buildings or Avatars. By default I used the schedule to change a setting
from true to false based on the INI the developers created. So, for
Building Damage True would ENABLE building damage. For Avatars True
would DISABLE avatars.  To make the scheduling easier I have allowed the
ability to flip the schedule. So, if you set FlipBuildingDmgSchedule=yes
then Building Damage will be DISABLED if it was currently 0600 and the
schedule was set to 1-0500to1-0700 .  This will allow for easier
scheduling based on server admin preferences.
Also, I added an option in the INI to turn off overwriting the default
INI. This is because some people reported it was not needed. However,
during testing on my server I had to remote the default INI to prevent
issues.