Modified to Default to Re-download Mods
This update changes the default action to redownload the mods to
automatically fill out modlist.txt . This is the much easier option as
long as the user had bandwidth available.

Also will now backup the old modlist file every time, and make sure
files are overwritten if exisiting. Finally, I also changed the modlist
file name to reflect that it is not temporary.