Sort mods by specific order & Download SteamCMD if needed
This is the final release for automatic mod sorting. This will allow the
mods to be sorted by the order listed in the INI file.  It creates a
seperate INI file to store the modID to modname conversion so this can
be accomplished. Also, this includes the update to automtically download
SteamCMD if it is not found in the directory specified.