Fix Server Password Bug
This fixes a bug introduced by last Conan Exiles update. If
ServerPassword is in the command line, even if it is blank, it shows as
a password protected server. I just removed the ServerPassword= and
AdminPassword= from the command line. These will be soley set by
modifying the INI files directly.